Add protection to file packets  that only nodes can read the packet

+++ node/node-base/src/main/java/net/suqatri/cloud/node/file/packet/FileDeletePacket.java
@@ -1,7 +1,9 @@
 package net.suqatri.cloud.node.file.packet;
 
 import lombok.Data;
+import net.suqatri.cloud.api.CloudAPI;
 import net.suqatri.cloud.api.impl.packet.CloudPacket;
+import net.suqatri.cloud.api.utils.ApplicationType;
 
 import java.io.File;
 
@@ -12,6 +14,7 @@ public class FileDeletePacket extends CloudPacket {
 
     @Override
     public void receive() {
+        if(CloudAPI.getInstance().getApplicationType() != ApplicationType.NODE) return;
         File file = new File(this.path);
         if(file.exists()) file.delete();
     }
